moved the ModelProcessor interface to general, as it is used by several packages
